testes usando netgen.c, correcoes em dimacs2nr.py

+++ src/nr/lubksb.c
@@ -1,4 +1,4 @@
-void lubksb(float **a, int n, int *indx, float b[])
+void lubksb(double **a, int n, int *indx, double b[])
 /*Solves the set of n linear equations A·X = B. Here a[1..n][1..n] is input, not as the matrix
 A but rather as its LU decomposition, determined by the routine ludcmp. indx[1..n] is input
 as the permutation vector returned by ludcmp. b[1..n] is input as the right-hand side vector
@@ -8,7 +8,7 @@ into account the possibility that b will begin with many zero elements, so it is
 in matrix inversion.*/
 {
 	int i,ii=0,ip,j;
-	float sum;
+	double sum;
 	for (i=1;i<=n;i++) { /*When ii is set to a positive value, it will become the
 		index of the ﬁrst nonvanishing element of b. We now
 		do the forward substitution, equation (2.3.6). The
