+++ src/nr/ludcmp.h
@@ -4,4 +4,4 @@ indx[1..n] is an output vector that records the row permutation eﬀected by the
 pivoting; d is output as ±1 depending on whether the number of row interchanges was even
 or odd, respectively. This routine is used in combination with lubksb to solve linear equations
 or invert a matrix.*/
-void ludcmp(float **a, int n, int *indx, float *d);
+void ludcmp(double **a, int n, int *indx, double *d);
